TestUtil: Move assertThrows(...) to Assert
TestUtil contains assertThrows(...) and
getFilePathInSandboxFolder(String). Both methods are doing things that
are unrelated to each other.

This makes TestUtil a class with low cohesion.

Let’s create a new class Assert and move assertThrows(...) into it.
