+++ src/test/java/seedu/address/model/UnmodifiableObservableListTest.java
@@ -1,7 +1,7 @@
 package seedu.address.model;
 
 import static org.junit.Assert.assertSame;
-import static seedu.address.testutil.TestUtil.assertThrows;
+import static seedu.address.testutil.Assert.assertThrows;
 
 import java.util.ArrayList;
 import java.util.Comparator;
