+++ src/test/java/seedu/address/testutil/Assert.java
@@ -1,0 +1,23 @@
+package seedu.address.testutil;
+
+import junit.framework.AssertionFailedError;
+
+/**
+ * A set of assertion methods useful for writing tests.
+ */
+public class Assert {
+    public static void assertThrows(Class<? extends Throwable> expected, Runnable executable) {
+        try {
+            executable.run();
+        } catch (Throwable actualException) {
+            if (actualException.getClass().isAssignableFrom(expected)) {
+                return;
+            }
+            String message = String.format("Expected thrown: %s, actual: %s", expected.getName(),
+                    actualException.getClass().getName());
+            throw new AssertionFailedError(message);
+        }
+        throw new AssertionFailedError(
+                String.format("Expected %s to be thrown, but nothing was thrown.", expected.getName()));
+    }
+}
